+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 * [#179 BUG: Text fields spotted in the wild, that corrupts export](https://github.com/WebTools-NG/WebTools-NG/issues/179)
 * [#103 AutoUpdate](https://github.com/WebTools-NG/WebTools-NG/issues/103)
-
+* [#21 ET Fields selection for custom levels](https://github.com/WebTools-NG/WebTools-NG/issues/21)
 
 ## V0.1.15